NotificationService/requirements.txt
SQLAlchemy>=2.0,<2.1
asyncpg>=0.29
aio-pika>=9.4.0
aiohttp>=3.9.0
python-dotenv==1.0.1
